firmware: improve wave storage
Wave samples are in the range [0x0,0xf], but were stored in a full byte.
Improve storage by storing two levels of volume per byte. Even volume
numbers are stored in the 4 LSBs or a byte, while odd volume numbers are
in the 4 MSBs. This allows us to reduce the amount of memory required to
store the wave samples by half.

Before: 1712 bytes (83%) of dynamic memory used.
After:  1200 bytes (58%) of dynamic memory used.