firmware: optimize mixing of chords/notes
If a channel's volume is zero, we can just skip it in the mixing routine
and same some CPU time. It was not desirable to do this before because
the rythm/arpeggio were very dependent on the global speed of the
program and doing so would increase the variation in rythm. But now that
rythm is updated according to real time we can apply this optimization.