firmware: use real time for rythm/arpeggio
The counters for rythm and arpeggio were updated in the main loop.
However since that loop has a variable execution time, the rythm could
go slower or faster depending on the load. This was especially
noticeable when playing chords and notes at the same time.

This patch updates the time-sensitive counters in the timer interrupt
handler, which we know will be called with absolute regularity. It also
refactors the code a bit so that rythm and arpeggio updates now have
their own functions. To avoid race conditions between the main loop and
the interrupt timer, interrupts are temporarily disabled when accessing
the shared counter variables.